refactor(settings/ai-providers): extract connect forms
Move ConnectApiKeyForm and ConnectOpenAICompatibleForm (plus their Zod
schemas and type aliases) into a new ai-providers/connect-forms.tsx file
so the forms can be reused by the upcoming ConnectProviderDialog.
